Fixed CreateEvent() and added Enum for Job,Class and DPSType
^

+++ maple-syrup-api/Services/Service/RequirementService.cs
@@ -18,13 +18,13 @@ namespace maple_syrup_api.Services.Service
         public int AddPlayer(Player player, EventRequirement pRequirement)
         {
             
-            if (pRequirement.ClassRequirement[player.Class] > 0)
+            if (pRequirement.ClassRequirement[(int) player.Class] > 0)
             {
                 //A spot if left. So we will check if other requirements are also met
                 if (!pRequirement.PreciseJob)
                 {
                     //We do not require only certain job
-                    pRequirement.ClassRequirement[player.Class]--;
+                    pRequirement.ClassRequirement[(int) player.Class]--;
                     pRequirement.PlayerCount++;
                     //ADD PLAYER TO REQUIREMENT AND SAVE REQUIREMENT/EVENT
                     return 0;
@@ -33,17 +33,17 @@ namespace maple_syrup_api.Services.Service
                 else
                 {
                     //If we do require OnePerJob
-                    if(pRequirement.PerJobRequirement[player.Job] >= 1)
+                    if(pRequirement.PerJobRequirement[(int) player.Job] >= 1)
                     {
-                        if((pRequirement.DPSRequiredByType) && (player.Class == 2))
+                        if((pRequirement.DPSRequiredByType) && ((int) player.Class == 2))
                         {
                             //If is a DPS and we have RequiredByType, then have to check if can add
 
                             if (pRequirement.DPSTypeRequirement[(int) player.DPSType] >= 1)
                             {
                                 //Then we all set
-                                pRequirement.PerJobRequirement[player.Job]--;
-                                pRequirement.ClassRequirement[player.Class]--;
+                                pRequirement.PerJobRequirement[(int) player.Job]--;
+                                pRequirement.ClassRequirement[(int) player.Class]--;
                                 pRequirement.DPSTypeRequirement[(int) player.DPSType]--;
                                 pRequirement.PlayerCount++;
                                 return 0;
@@ -57,8 +57,8 @@ namespace maple_syrup_api.Services.Service
                         }
                         else
                         {
-                            pRequirement.PerJobRequirement[player.Job]--;
-                            pRequirement.ClassRequirement[player.Class]--;
+                            pRequirement.PerJobRequirement[(int) player.Job]--;
+                            pRequirement.ClassRequirement[(int) player.Class]--;
                             pRequirement.PlayerCount++;
                             return 0;
                         }
@@ -89,9 +89,9 @@ namespace maple_syrup_api.Services.Service
 
                     Player removedPlayer = pRequirement.Players[i];
 
-                    pRequirement.ClassRequirement[removedPlayer.Class]++;//Add a new spot for this class
-                    if (pRequirement.PreciseJob) pRequirement.PerJobRequirement[removedPlayer.Job]++;//Add a new spot for this job if PreciseJob is true
-                    if (pRequirement.DPSRequiredByType && (removedPlayer.Class == 2)) pRequirement.DPSTypeRequirement[(int) removedPlayer.DPSType]++;//Add a new spot for DPSType if DPSType is true and player is a DPS
+                    pRequirement.ClassRequirement[(int) removedPlayer.Class]++;//Add a new spot for this class
+                    if (pRequirement.PreciseJob) pRequirement.PerJobRequirement[(int) removedPlayer.Job]++;//Add a new spot for this job if PreciseJob is true
+                    if (pRequirement.DPSRequiredByType && ((int) removedPlayer.Class == 2)) pRequirement.DPSTypeRequirement[(int) removedPlayer.DPSType]++;//Add a new spot for DPSType if DPSType is true and player is a DPS
 
                     pRequirement.Players[i] = null;//removes player object from list
                     pRequirement.PlayerCount--;
